Some code refactor to isolate code.
